ai tutor 2/requirements.txt
pdfplumber
pinecone
requests
python-dotenv
google-generativeai
fastapi
uvicorn[standard]
python-multipart